requirements.txt
fastapi
pydantic
uvicorn
gunicorn
aiologger[aiofiles]
requests
openai
google-api-python-client
python-dotenv
beautifulsoup4

# serpapi (если удобно использовать)
# beautifulsoup4 (для парсинга страниц и т.д.)